Bulletin Board, Whiteboards, Bills
This is all of the Java code and XML that I had to write to get the
Bulletin Board, Whiteboards, and Bills functions to work.

I updated the Android Manifest with all of the necessary activities.

Added additional constructors to the Bulletin Board Post object in
accordance with Firebase's needs.

Corrected a typo in commented out code in the House class.

Uncommented configure methods and calls to the Bulletin Board,
Whiteboard, and Bills features in the Main Activity.

Updated the message class to implement Serializable.

Added a method to the PickHouseActivity that should generate (and
overwrite old) base Whiteboards between that user and all previous
users.

Added a toString method to the User class for debugging purposes.